+++ src/Accounts/Authentication/Properties/Resources.Designer.cs
@@ -277,11 +277,29 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to When enabled, Azure PowerShell will check for updates automatically and display a hint message when an update is available. The default value will be changed from false to true in Az version 11.0.0..
+        /// </summary>
+        public static string HelpMessageOfCheckForUpgrade {
+            get {
+                return ResourceManager.GetString("HelpMessageOfCheckForUpgrade", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Subscription name or GUID. Sets the default context for Azure PowerShell when logging in without specifying a subscription..
         /// </summary>
         public static string HelpMessageOfDefaultSubscriptionForLogin {
             get {
                 return ResourceManager.GetString("HelpMessageOfDefaultSubscriptionForLogin", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to When disabled, error records will not be written to ~/.Azure/ErrorRecords. This config will be replaced by &quot;EnableErrorRecordsPersistence&quot; as opt-in in the next major release of Az around November 2023..
+        /// </summary>
+        public static string HelpMessageOfDisableErrorRecordsPersistence {
+            get {
+                return ResourceManager.GetString("HelpMessageOfDisableErrorRecordsPersistence", resourceCulture);
             }
         }
         
@@ -309,6 +327,15 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Properties {
         public static string HelpMessageOfEnableDataCollection {
             get {
                 return ResourceManager.GetString("HelpMessageOfEnableDataCollection", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to When enabled, error records will be written to ~/.Azure/ErrorRecords..
+        /// </summary>
+        public static string HelpMessageOfEnableErrorRecordsPersistence {
+            get {
+                return ResourceManager.GetString("HelpMessageOfEnableErrorRecordsPersistence", resourceCulture);
             }
         }