+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/NewAzManagementGroupDeploymentStack.cs
@@ -100,12 +100,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
             {
                 Hashtable parameters = new Hashtable();
                 string filePath = "";
-                string parameterFilePath = "";
-
-                if (BicepUtility.IsBicepparamFile(TemplateParameterFile) && !BicepUtility.IsBicepFile(TemplateFile))
-                {
-                    throw new NotSupportedException($"Bicepparam file {TemplateParameterFile} is only supported with a Bicep template file");
-                }
 
                 switch (ParameterSetName)
                 {
@@ -122,13 +116,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         break;
                     case ParameterFileTemplateSpecParameterSetName:
                     case ParameterFileTemplateUriParameterSetName:
-                        parameterFilePath = this.TryResolvePath(TemplateParameterFile);
-                        if (!File.Exists(parameterFilePath))
-                        {
-                            throw new PSInvalidOperationException(
-                                string.Format(ProjectResources.InvalidFilePath, TemplateParameterFile));
-                        }
-                        parameters = this.GetParameterObject(parameterFilePath);
+                        parameters = ResolveParameters();
 
                         // contruct the protected template URI if a query string was provided
                         if (!string.IsNullOrEmpty(QueryString))
@@ -140,6 +128,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         }
                         break;
                     case ParameterFileTemplateFileParameterSetName:
+                        parameters = ResolveParameters();
+
                         filePath = this.TryResolvePath(TemplateFile);
                         if (!File.Exists(filePath))
                         {
@@ -148,14 +138,10 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         }
                         filePath = ResolveBicepFile(filePath);
 
-                        parameterFilePath = this.TryResolvePath(TemplateParameterFile);
-                        if (!File.Exists(parameterFilePath))
-                        {
-                            throw new PSInvalidOperationException(
-                                string.Format(ProjectResources.InvalidFilePath, TemplateParameterFile));
-                        }
-                        parameters = this.GetParameterObject(ResolveBicepParameterFile(parameterFilePath));
                         TemplateUri = filePath;
+                        break;
+                    case ByParameterFileWithNoTemplateParameterSetName:
+                        parameters = ResolveParameters();
                         break;
                     case ParameterObjectTemplateFileParameterSetName:
                         filePath = this.TryResolvePath(TemplateFile);
@@ -213,6 +199,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         location: Location,
                         templateUri: !string.IsNullOrEmpty(protectedTemplateUri) ? protectedTemplateUri : TemplateUri,
                         templateSpec: TemplateSpecId,
+                        templateJson: TemplateJson,
                         parameterUri: TemplateParameterUri,
                         parameters: parameters,
                         description: Description,
@@ -230,7 +217,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                     WriteObject(deploymentStack);
                 };
 
-                if (!Force.IsPresent && currentStack == null)
+                if (!Force.IsPresent && currentStack != null)
                 {
                     string confirmationMessage = $"The DeploymentStack '{Name}' you're trying to create already exists in ManagementGroup '{ManagementGroupId}'. " +
                         $"Do you want to overwrite it\n?" +