+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/NewAzResourceGroupDeploymentStack.cs
@@ -92,12 +92,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
             {
                 Hashtable parameters = new Hashtable();
                 string filePath = "";
-                string parameterFilePath = "";
-
-                if (BicepUtility.IsBicepparamFile(TemplateParameterFile) && !BicepUtility.IsBicepFile(TemplateFile))
-                {
-                    throw new NotSupportedException($"Bicepparam file {TemplateParameterFile} is only supported with a Bicep template file");
-                }
 
                 switch (ParameterSetName)
                 {
@@ -114,13 +108,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         break;
                     case ParameterFileTemplateSpecParameterSetName:
                     case ParameterFileTemplateUriParameterSetName:
-                        parameterFilePath = this.TryResolvePath(TemplateParameterFile);
-                        if (!File.Exists(parameterFilePath))
-                        {
-                            throw new PSInvalidOperationException(
-                                string.Format(ProjectResources.InvalidFilePath, TemplateParameterFile));
-                        }
-                        parameters = this.GetParameterObject(parameterFilePath);
+                        parameters = ResolveParameters();
 
                         // contruct the protected template URI if a query string was provided
                         if (!string.IsNullOrEmpty(QueryString))
@@ -132,6 +120,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         }
                         break;
                     case ParameterFileTemplateFileParameterSetName:
+                        parameters = ResolveParameters();
+
                         filePath = this.TryResolvePath(TemplateFile);
                         if (!File.Exists(filePath))
                         {
@@ -140,14 +130,10 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         }
                         filePath = ResolveBicepFile(filePath);
 
-                        parameterFilePath = this.TryResolvePath(TemplateParameterFile);
-                        if (!File.Exists(parameterFilePath))
-                        {
-                            throw new PSInvalidOperationException(
-                                string.Format(ProjectResources.InvalidFilePath, TemplateParameterFile));
-                        }
-                        parameters = this.GetParameterObject(ResolveBicepParameterFile(parameterFilePath));
                         TemplateUri = filePath;
+                        break;
+                    case ByParameterFileWithNoTemplateParameterSetName:
+                        parameters = ResolveParameters();
                         break;
                     case ParameterObjectTemplateFileParameterSetName:
                         filePath = this.TryResolvePath(TemplateFile);
@@ -201,6 +187,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         resourceGroupName: ResourceGroupName,
                         templateUri: !string.IsNullOrEmpty(protectedTemplateUri) ? protectedTemplateUri : TemplateUri,
                         templateSpec: TemplateSpecId,
+                        templateJson: TemplateJson,
                         parameterUri: TemplateParameterUri,
                         parameters: parameters,
                         description: Description,
@@ -217,7 +204,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                     WriteObject(deploymentStack);
                 };
 
-                if (!Force.IsPresent && currentStack == null)
+                if (!Force.IsPresent && currentStack != null)
                 {
                     string confirmationMessage = $"The DeploymentStack '{Name}' you're trying to create already exists in ResourceGroup '{ResourceGroupName}'. " +
                         $"Do you want to overwrite it?\n" +