+++ src/ContainerInstance/ChangeLog.md
@@ -20,6 +20,9 @@
 
 ## Upcoming Release
 
+## Version 3.2.1
+* Fixed a bug in `Invoke-AzContainerInstanceCommand` when no result was returned under some conditions [#22453]
+
 ## Version 3.2.0
 * Added `priority` property to Container Group properties
 * Added `Confidential` sku type to Container Group Skus