+++ src/ContainerRegistry/ContainerRegistry/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 4.1.1
+* Updated Azure.Core to 1.34.0.
+
 ## Version 4.1.0
 * Updated Azure.Core to 1.33.0.
 * Added new cmdlet `New-AzContainerRegistryCredentials`