+++ src/Maintenance/Maintenance/ChangeLog.md
@@ -20,6 +20,9 @@
 
 ## Upcoming Release
 
+## Version 1.3.1
+* Fixed breaking change information
+  
 ## Version 1.3.0
 * Added support for Resource Group and Subscription configuration assignment.