+++ src/SecurityInsights/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 
 ## Upcoming Release
+* Fixed parameters' issues for `New-AzSentinelAlertRule` and `Update-AzSentinelAlertRule` [#21181][#21217][#22318]
+
+## Version 3.0.2
+* Added breaking change message for `Az.SecurityInsights`.
 
 ## Version 3.0.1
 * Fixed for `Update-AzSentinelAlertRule` fails when using `-TriggerThreshold 0` [#20417]